combine pipelines, utils #178
