+++ scripts/vandv/graphs.py
@@ -15,7 +15,7 @@ from scipy.stats import trim_mean
 from sklearn.metrics import mean_squared_error, mean_absolute_error
 from tqdm import tqdm
 
-from scripts.utils.Utils_emtech import iso_to_gregorian
+from scripts.utils.utils import iso_to_gregorian
 
 
 def trim_leading_zero_counts(dates, values):
